requirements.txt
Django==4.2.6
django-dry-tests==1.0.0
django-find-similar==1.3.0
pandas==2.1.1
openpyxl==3.1.2
coverage==7.3.2
mixer==7.2.2
django-debug-toolbar==4.2.0
pylint==3.0.2